quartercircle calc namespace refactor in ui
