+++ ExperimentalProbability.UI/MVVM/Views/Calculations/TwoDPlane/Pi/DescriptionView.xaml.cs
@@ -1,6 +1,6 @@
 ﻿using System.Windows.Controls;
 
-namespace ExperimentalProbability.UI.MVVM.Views.Calculations.TwoDPlane.QuarterCircle
+namespace ExperimentalProbability.UI.MVVM.Views.Calculations.TwoDPlane.Pi
 {
     public partial class DescriptionView : UserControl
     {
