+++ ExperimentalProbability.UI/MVVM/Views/Calculations/TwoDPlane/Pi/ShellView.xaml
@@ -1,11 +1,11 @@
 ﻿<UserControl
-    x:Class="ExperimentalProbability.UI.MVVM.Views.Calculations.TwoDPlane.QuarterCircle.ShellView"
+    x:Class="ExperimentalProbability.UI.MVVM.Views.Calculations.TwoDPlane.Pi.ShellView"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:cal="http://www.caliburnproject.org"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:vm="clr-namespace:ExperimentalProbability.UI.MVVM.ViewModels.Calculations.TwoDPlane.QuarterCircle"
+    xmlns:vm="clr-namespace:ExperimentalProbability.UI.MVVM.ViewModels.Calculations.TwoDPlane.Pi"
     Padding="5"
     d:DataContext="{d:DesignInstance Type=vm:ShellViewModel}"
     mc:Ignorable="d"
