+++ ExperimentalProbability.UI/MVVM/Views/Calculations/TwoDPlane/Pi/ShellView.xaml.cs
@@ -1,6 +1,6 @@
 ﻿using System.Windows.Controls;
 
-namespace ExperimentalProbability.UI.MVVM.Views.Calculations.TwoDPlane.QuarterCircle
+namespace ExperimentalProbability.UI.MVVM.Views.Calculations.TwoDPlane.Pi
 {
     public partial class ShellView : UserControl
     {
